backend/requirements.txt
bcrypt==3.2.0
email-validator==1.1.2
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
numpy==1.21.6
packaging==20.9
pandas==1.3.5
psycopg2-binary==2.9.4
py==1.10.0
pydantic==1.10.2
PyJWT==2.0.1
pytest==6.2.2
python-dotenv==0.15.0
SQLAlchemy==1.4.3
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.4
uvicorn==0.13.4
